requirements.txt
# CyberWatch — Full Requirements
# pip install -r requirements.txt

# Core data + ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Dashboard
streamlit>=1.32.0
plotly>=5.18.0

# Big data (optional — only needed for spark/spark_processor.py)
# pyspark>=3.5.0

# Agent frameworks (optional upgrade — agents work without these)
# langchain>=0.1.0
# langchain-community>=0.0.20
# crewai>=0.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0
watchdog>=3.0.0
